Randomly shuffle block locations at the Driver.
Even if nodes are grouped by node, rack, and remote order,
the order within these groups remain the same.
This can lead to certain nodes becoming overloaded.
Thus randomized shuffling is applied within locality groups.
This is done at the Driver, because it has visibility of the group
that each node belongs to (and the Client does not).