Workaround to use HdfsBlockId when write data
As the metadata used in HDFS do not exist until the blocks
are written to the HDFS. To avoid the meaningless overhead,
the parameters are filled with fake values to generate BlockId and
equals() does not take into account these metadata.

By the way, the redundant parameters should be removed to get rid of
the vicious dependency on hadoop.