[MINOR] Try releasing worker asynchronously on worker delete (#720)
This PR fix the bug in SynchronizationManager.

Its onWorkerDeleted method is called by AsyncDolphinDriver when an worker context has been closed by EM's Delete.

In the current implementation, this method is a blocking call and it tries releasing workers synchronously when workers are ready to be released by this Delete event. But, this releasing try may be blocked if there's ongoing optimization by EM.

Then AsyncDolphinDriver, a caller of the method, will be blocked during handling the closed context during EM's Delete. As a result, it's a deadlock.

This PR fixes onWorkerDeleted to call tryReleaseWorkers asynchronously with a new thread.